refactored menu iconic for a better DX
- split MenuIconicItemLogin into two components based on the auth status
- fixed bug with missing menu iconic on medium-sized devices
